+++ examples/titanic/run_tests.py
@@ -18,7 +18,7 @@ from logging import StreamHandler, DEBUG
 
 from trilogy_nlp.constants import logger
 from trilogy_nlp.main import build_query
-from trilogy_nlp.core import NLPEngine
+from trilogy_nlp.llm_interface import NLPEngine
 
 # how many passengers survived in first and second class?
