+++ trilogy_nlp/environment.py
@@ -3,14 +3,12 @@ from langchain_core.language_models import BaseLanguageModel
 from trilogy import Environment
 from pydantic import BaseModel
 from langchain.agents import create_structured_chat_agent, AgentExecutor
-from trilogy.core.models import (
-    Environment,
-)
 from langchain.tools import Tool, StructuredTool
 from trilogy.core.enums import Purpose
 from langchain_core.prompts import ChatPromptTemplate, MessagesPlaceholder
 from trilogy_nlp.tools import get_wiki_tool
 from trilogy.core.enums import Modifier
+from difflib import get_close_matches
 
 # from trilogy.core.exceptions import Import
 
@@ -32,7 +30,11 @@ def get_environment_detailed_values(env: Environment, input: str):
     try:
         new.parse(f"import {input} as {input};")
     except ImportError:
-        return "This is an invalid database. Refer back to the list from the get_database_list tool."
+        suggestions = get_close_matches(input, get_environment_possible_imports(env))
+        base = "This is an invalid database. Refer back to the list from the get_database_list tool. Names must match exactly."
+        if suggestions:
+            base += f" Did you mean {suggestions[0]}?"
+        return base
     # TODO: does including a description work better?
     return {
         k.split(".", 1)[1]
@@ -240,6 +242,10 @@ def build_env_and_imports(
             continue
         # only merge on keys automatically
         if not concept.purpose == Purpose.KEY:
+            continue
+        # don't merge on date
+        # TODO: be less hacky in terms of what default merges we should do
+        if "date" in concept.namespace:
             continue
         pre_namespace_root = k.split(".", 1)[1]
         if "." not in pre_namespace_root:
