+++ trilogy_nlp/llm_interface/constants.py
@@ -1,0 +1,6 @@
+# llm will not know how to use these; better not to include
+COMPLICATED_FUNCTIONS = ["ALIAS", "CUSTOM", "GROUP", "UNNEST", "ATTR_ACCESS"]
+
+
+
+MAGIC_GENAI_DESCRIPTION = 'local to select'
